backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.5.0
django-environ==0.11.2
django-filter==24.3
django-storages==1.14.4
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.8
flower==2.0.1
frozenlist==1.5.0
googlemaps==4.10.0
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
multidict==6.1.0
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==11.3.0
tornado==6.4.1
twilio==9.3.7
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.0
